# src/docs/swagger.yml
swagger: '2.0'
info:
  description: API Orders Service
  version: 1.0.0
  title: API Orders Service
  termsOfService: ''
  contact:
    email: caotri1242001@gmail.com
host: api.127.0.0.1.xip.ip
basePath: /api/v1
...
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /success:
    get:
      x-swagger-router-controller: order
      operationId: success
      description: success Paypal
      tags:
        - ERP-order
      parameters:
        - name: PayerID
          type: string
          in: query
          required: false
        - name: paymentId
          type: string
          in: query
          required: false
      responses:
        '200':
          description: Pong
          schema:
            type: object
            items:
              $ref: '#/definitions/Pong'
...
definitions:
  Pong:
    type: object
...